// src/modules/products/interfaces/product.interface.ts
import type { User } from '@/modules/auth/interfaces';

export interface Product {
  id: string;
  title: string;
  price: number;
  description: string;
  slug: string;
  stock: number;
  sizes: Size[];
  gender: string;
  isActive: boolean;
  tags: string[];
  images: string[];
  user: User;
}

export enum Size {
  XS = 'XS',
  S = 'S',
  M = 'M',
  L = 'L',
  XL = 'XL',
  XXL = 'XXL',
}